+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTableCellRenderer.java
@@ -61,7 +61,7 @@ public final class EntityTableCellRenderer {
 		FilterTableCellRenderer.Builder<Entity, Attribute<?>, T> builder = FilterTableCellRenderer.builder()
 						.<Entity, Attribute<?>, T>columnClass(attributeDefinition.attribute().type().valueClass())
 						.uiSettings(new EntityUISettings(queryCondition))
-						.string(attributeDefinition::string);
+						.formatter(attributeDefinition::string);
 		if (!attributeDefinition.items().isEmpty()) {
 			// Otherwise the horizontal aligment is based on the Item value type
 			builder.horizontalAlignment(FilterTableCellRenderer.HORIZONTAL_ALIGNMENT.getOrThrow());