+++ tools/loadtest/ui/src/main/java/is/codion/tools/loadtest/ui/LoadTestPanel.java
@@ -417,7 +417,7 @@ public final class LoadTestPanel<T> extends JPanel {
 						.scrollToSelectedItem(false)
 						.cellRenderer(ApplicationRow.DURATION, FilterTableCellRenderer.builder()
 										.columnClass(Integer.class)
-										.string(duration -> duration == null ? "" : DURATION_FORMAT.format(duration))
+										.formatter(duration -> duration == null ? "" : DURATION_FORMAT.format(duration))
 										.build())
 						.popupMenuControls(table -> Controls.builder()
 										.control(Control.builder()